requirements.txt
# Core dependencies
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.24.0

# Embeddings and Vector Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Utilities
pathlib>=1.0.1
